website/requirements.txt
Click==7.0
Faker==2.0.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.2
Werkzeug==0.15.6
